refactor: move VhostUserVsockThread into a different file
